Moved builder classes into the same namespace of the things they build
