+++ src/Framework/src/Routing/Components/RouterComponent.php
@@ -20,10 +20,10 @@ use Aphiria\DependencyInjection\ResolutionException;
 use Aphiria\DependencyInjection\TargetedContext;
 use Aphiria\Net\Http\IRequestHandler;
 use Aphiria\Routing\Attributes\AttributeRouteRegistrant;
-use Aphiria\Routing\Builders\RouteCollectionBuilder;
-use Aphiria\Routing\Builders\RouteCollectionBuilderRouteRegistrant;
 use Aphiria\Routing\IRouteRegistrant;
 use Aphiria\Routing\RouteCollection;
+use Aphiria\Routing\RouteCollectionBuilder;
+use Aphiria\Routing\RouteCollectionBuilderRouteRegistrant;
 use Aphiria\Routing\RouteRegistrantCollection;
 use Closure;
 use RuntimeException;
