+++ src/Framework/tests/Routing/Components/RouterComponentTest.php
@@ -15,9 +15,9 @@ namespace Aphiria\Framework\Tests\Routing\Components;
 use Aphiria\DependencyInjection\Container;
 use Aphiria\Framework\Routing\Components\RouterComponent;
 use Aphiria\Routing\Attributes\AttributeRouteRegistrant;
-use Aphiria\Routing\Builders\RouteBuilder;
-use Aphiria\Routing\Builders\RouteCollectionBuilder;
+use Aphiria\Routing\RouteBuilder;
 use Aphiria\Routing\RouteCollection;
+use Aphiria\Routing\RouteCollectionBuilder;
 use Aphiria\Routing\RouteRegistrantCollection;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
