+++ src/Router/src/Attributes/AttributeRouteRegistrant.php
@@ -17,12 +17,12 @@ use Aphiria\Middleware\Attributes\Middleware as MiddlewareLibraryMiddleware;
 use Aphiria\Reflection\ITypeFinder;
 use Aphiria\Reflection\TypeFinder;
 use Aphiria\Routing\Attributes\Controller as ControllerAttribute;
-use Aphiria\Routing\Builders\RouteCollectionBuilder;
-use Aphiria\Routing\Builders\RouteGroupOptions;
 use Aphiria\Routing\IRouteRegistrant;
 use Aphiria\Routing\Matchers\Constraints\IRouteConstraint;
 use Aphiria\Routing\Middleware\MiddlewareBinding;
 use Aphiria\Routing\RouteCollection;
+use Aphiria\Routing\RouteCollectionBuilder;
+use Aphiria\Routing\RouteGroupOptions;
 use ReflectionAttribute;
 use ReflectionClass;
 use ReflectionException;
